requirements.txt
boto3
confluent-kafka
prefect
psycopg2-binary
pyspark
pytest
